Update Day11_2D_Arrays
Here I used a nested for loop to run through the given hourglass patterns, comparing the sum to the current largest sum and replacing it as needed. The hardest part came when the test cases started introducing negative values into the array. The solution I ended up with was to hard-code the starting value of largest as the lowest possible sum of an hourglass based on the constraints given (-9 <= n <= 9).