Update Day13_Abstract_Classes
The instructions were very clear for this challenge. I did struggle at first with the fact that a derived class makes an implicit call to the parent class, and that call has to match an existing constructor, but after figuring out what the error message was telling me, writing code that created and object and printed its contents was easy enough.