Update Day17_More_Exceptions
My solution checks that n and p are both positive and if they are it multiplies n p times. If at least one of them isn’t it throws the required exception. 
Note: I used a for loop to manage the math. Math.Pow(n, p) would have been a neater solution.