Update Day18_Queues_and_Stacks
The instructions lay out the desired solution clearly, and a lot of the code is provided. I created a stack and a queue and the required methods for the stub code to call and compare the results. I’ve seen a simpler method for checking for palindromes involving treating the string like an array and comparing s[i] to s[n-i]-1 (where i is a loop counter and n is s.Length)until I reaches n/2, but without being able to edit the stub code I wasn’t able to implement it.