Update Day19_Interfaces
The divisorSum() method checks each number between 1 and the square root of n for possible divisors. If it finds a number that divides evenly into n (by checking n % i) it runs a second check. We don’t want to count the square root of n twice, so we put that in its own if statement. The other divisors are accumulated and returned as n.