Update Day20_Sorting
Here the challenge calls for a simple bubble sort (a nested pair of loops that rearrange the contents of a data set by comparing 2 elements and swapping them if they aren’t in order). The sort goes through each pair and exchanges them based on the if statement. If it doesn’t find any swaps it breaks early.