Update Day22_Binary_Search_Trees
The provided code produces a binary search tree. I wrote two while loops that get the depth of the tree. The first loop searches the right side of the tree and returns the depth, the second loop does the same for the left side. The two values are compared and the largest one is the depth of the tree.