Update Day23_BST_Level-Order_Traversal
Like the previous challenge the stub code produces a binary search tree. This time I used a while loop and a queue to print the tree in breadth-first order. Root is assigned as current, removed from the queue and printed. Then its children are added to the queue, the first one is removed, printed, and it’s children are added to the queue before the pointer goes to the second child. This repeats until only the leaves remain in the queue and they get dequeued and printed.