Update Day24_More_Linked_Lists
The while loop in the removeDuplicates() method looks at each node in the list and compares it’s data to the next node’s data. If they match it assigns the node after next to the next position and checks again, otherwise it moves the pointer down the list and checks for a duplicate. After each node has been checked it returns the pointer to the head of the list.