Update Day26_Nested_Logic
The Main() function takes the inputs from the user and splits the strings int string arrays. It then calls the CalculateFine() function which parses the arrays into the constituent parts of the date. As its written the code doesn’t account for the dates being formatted the wrong way, and the function of parsing the strings could be separated into another function. The parts of the dates are compared to determine the correct calculation to apply and the value of the fine is returned and printed.