Update Day28_RegEx,_Patterns,_and_Intro_to_Databases
The challenge was to learn regular expressions and use them to sort through the given data. Regular expressions seem like a powerful tool that I would be interested in learning about but I opted to save that for a later day. My solution instead was to go through each piece of input as it came in and accumulate a list of Gmail accounts, sort it, and then print it.